refactor the setup/teardown stuff from the tests.
