+++ tests/test_feeds.py
@@ -7,43 +7,24 @@ from os.path import abspath, dirname
 from lxml import etree as ET
 
 sys.path.append(dirname(dirname(abspath(__file__))))
+from tests.pln_testcase import PkpPlnTestCase
 from webapp.feeds.feed_server import FeedsApp
 import pkppln
 
 parser = ET.XMLParser()
 
 
-class TestFeeds(unittest.TestCase):
+class TestFeeds(PkpPlnTestCase):
 
     @classmethod
     def setUpClass(self):
         self.app = FeedsApp("Feeds")
-        self.handle = pkppln.get_connection()
+        super(TestFeeds, self).setUpClass()
 
     @classmethod
     def tearDownClass(self):
-        pkppln.db_execute('TRUNCATE TABLE terms_of_use', db=self.handle)
-        self.handle.commit()
+        super(TestFeeds, self).tearDownClass()
         self.app = None
-
-    def setUp(self):
-        pkppln.db_execute('TRUNCATE TABLE terms_of_use', db=self.handle)
-        sql = """
-INSERT INTO terms_of_use (weight, key_code, lang_code, content)
-VALUES (%s, %s, %s, %s)
-"""
-        data = [
-            (0, 'utf8.single', 'en-US', u'I am good to go.'),
-            (1, 'utf8.double', 'en-US', u'U+00E9: \u00E9'),
-            (2, 'utf8.triple', 'en-US', u'U+20AC: \u20AC'),
-            (3, 'typographic.doublequote', 'en-US',
-             u'U+201C U+201D: \u201C \u201D'),
-            (4, 'single.anglequote', 'en-US', u'U+2039 U+203A: \u2039 \u203A'),
-            (0, 'utf8.single', 'en-CA', u'I am good to go Canada'),
-            (1, 'utf8.double', 'en-CA', u'U+00E9: \u00E9 Canada'),
-        ]
-        self.handle.cursor().executemany(sql, data)
-        self.handle.commit()
 
     def test_setup(self):
         self.assertIsInstance(self.app, FeedsApp, "Saved a feeds app for use.")
